requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==3.0.1
alembic==1.13.1
openpyxl==3.1.2
openai==2.14.0
pandas==2.1.4
requests==2.32.3


